Split ddo.py to encrypt.py/decrypt.py. Move errors to error_responses.py
