+++ ocean_provider/routes/decrypt.py
@@ -25,8 +25,9 @@ from ocean_provider.utils.data_nft import (
     get_metadata,
     get_metadata_logs_from_tx_receipt,
 )
-from ocean_provider.utils.encryption import do_decrypt, do_encrypt
-from ocean_provider.utils.util import get_request_data, service_unavailable
+from ocean_provider.utils.encryption import do_decrypt
+from ocean_provider.utils.error_responses import error_response, service_unavailable
+from ocean_provider.utils.util import get_request_data
 from ocean_provider.validation.provider_requests import DecryptRequest
 from web3.main import Web3
 
@@ -38,79 +39,6 @@ requests_session = get_requests_session()
 requests_session.mount("file://", LocalFileAdapter())
 
 logger = logging.getLogger(__name__)
-
-standard_headers = {"Content-type": "text/plain", "Connection": "close"}
-# {'charset': 'utf-8'}
-
-
-@services.route("/encryptDDO", methods=["POST"])
-def encryptDDO():
-    """Encrypt DDO using the Provider's own symmetric key (symmetric encryption).
-    This can be used by the publisher of an asset to encrypt the DDO of the
-    asset data files before publishing the asset DDO. The publisher to use this
-    service is one that is using a front-end with a wallet app such as MetaMask.
-    The DDO is encrypted by the provider so that the provider will be able
-    to decrypt at time of providing the service later on.
-
-    ---
-    tags:
-      - services
-    consumes:
-      - application/json
-    parameters:
-      - in: body
-        name: body
-        required: true
-        description: Asset urls encryption.
-        schema:
-          type: object
-          required:
-            - documentId
-            - document
-            - publisherAddress:
-          properties:
-            documentId:
-              description: Identifier of the asset to be registered in ocean.
-              type: string
-              example: 'did:op:08a429b8529856d59867503f8056903a680935a76950bb9649785cc97869a43d'
-            ddo:
-              description: document description object (DDO)
-              type: string
-              example: See https://github.com/oceanprotocol/docs/blob/feature/ddo_v4/content/concepts/did-ddo.md
-            publisherAddress:
-              description: Publisher address.
-              type: string
-              example: '0x00a329c0648769A73afAc7F9381E08FB43dBEA72'
-    responses:
-      201:
-        description: DDO successfully encrypted.
-      503:
-        description: Service Unavailable
-
-    return: the encrypted DDO (hex str)
-    """
-    if request.content_type != "application/octet-stream":
-        return error_response(
-            "Invalid request content type: should be application/octet-stream", 400
-        )
-
-    data = request.get_data()
-    logger.info(f"encryptDDO endpoint called. {data}")
-
-    try:
-        return _encryptDDO(data)
-    except Exception as e:
-        return service_unavailable(e, data, logger)
-
-
-def _encryptDDO(data: bytes) -> Response:
-
-    try:
-        encrypted_data = do_encrypt(data, provider_wallet)
-        logger.info(f"encrypted_data = {encrypted_data}")
-    except Exception:
-        return error_response(f"Failed to encrypt.", 400)
-    return Response(encrypted_data, 201, headers=standard_headers)
 
 
 @services.route("/decryptDDO", methods=["POST"])
@@ -155,7 +83,6 @@ def _decryptDDO(
     if authorized_decrypters and decrypter_address not in authorized_decrypters:
         return error_response(f"Decrypter not authorized", 403)
 
-    # Get arguments from transaction_id
     if transaction_id:
         try:
             tx_receipt = web3.eth.get_transaction_receipt(transaction_id)
@@ -166,7 +93,6 @@ def _decryptDDO(
                     "More than 1 MetadataCreated/MetadataUpdated event detected. "
                     "Using the event at index 0."
                 )
-
             log = logs[0]
             data_nft_address = log.address
             # Interpret "data" as utf-8 encoded bytes
@@ -192,10 +118,6 @@ def _decryptDDO(
             document_hash = Web3.toBytes(hexstr=document_hash)
         except Exception:
             return error_response(f"Failed converting input args to bytes.", 400)
-
-    assert isinstance(encrypted_document, bytes)
-    assert isinstance(flags, bytes)
-    assert isinstance(document_hash, bytes)
 
     # Check if DDO metadata state is ACTIVE
     (_, _, metadata_state, _) = get_metadata(web3, data_nft_address)
@@ -227,10 +149,6 @@ def _decryptDDO(
             "Document not encrypted (flags bit 2 not set). Skipping decryption."
         )
 
-    assert isinstance(encrypted_document, bytes)
-    assert isinstance(flags, bytes)
-    assert isinstance(document_hash, bytes)
-
     # bit 1:  check if DDO is lzma compressed
     if flags[0] & 1:
         try:
@@ -244,18 +162,11 @@ def _decryptDDO(
     document = working_document
     logger.info(f"document = {document}")
 
-    assert isinstance(encrypted_document, bytes)
-    assert isinstance(flags, bytes)
-    assert isinstance(document_hash, bytes)
-
     # Verify checksum matches
     if sha256(document).hexdigest() != document_hash.hex():
         return error_response("Checksum doesn't match.", 400)
     logger.info(f"Checksum matches.")
 
-    return Response(document, 201, standard_headers)
-
-
-def error_response(err_str: str, status: int) -> Response:
-    logger.error(err_str)
-    return Response(err_str, status, standard_headers)
+    return Response(
+        document, 201, {"Content-type": "text/plain", "Connection": "close"}
+    )
