+++ ocean_provider/utils/util.py
@@ -344,23 +344,6 @@ def decode_from_data(data, key, dec_type="list"):
     return data
 
 
-def service_unavailable(error, context, custom_logger=None):
-    text_items = []
-    for key, value in context.items():
-        value = value if isinstance(value, str) else json.dumps(value)
-        text_items.append(key + "=" + value)
-
-    logger_message = "Payload was: " + ",".join(text_items)
-    custom_logger = custom_logger if custom_logger else logger
-    custom_logger.error(logger_message, exc_info=1)
-
-    return Response(
-        json.dumps({"error": str(error), "context": context}),
-        503,
-        headers={"content-type": "application/json"},
-    )
-
-
 def check_asset_consumable(asset, consumer_address, logger, custom_url=None):
     code = asset.is_consumable({"type": "address", "value": consumer_address})
